+++ src/main/java/com/codegym/fashionshop/repository/bill/IBillRepository.java
@@ -16,13 +16,13 @@ public interface IBillRepository extends JpaRepository<Bill,Long> {
 
     @Query(value = "select sum(bi.quantity * p.price) as daily_revenue from bill_items as bi inner join bills as b on bi.bill_id = b.bill_id inner join pricings\n" +
             "as p on bi.pricing_id = p.pricing_id where date(b.date_create) = :date", nativeQuery = true)
-    Long getDailySalesRevenue(@Param("date") LocalDate date);
+    Double getDailySalesRevenue(@Param("date") LocalDate date);
     @Query(value = "select sum(bi.quantity * p.price) as monthly_revenue from bill_items as bi " +
             "inner join bills as b on bi.bill_id = b.bill_id " +
             "inner join pricings as p on bi.pricing_id = p.pricing_id " +
             "where year(b.date_create) = :year and month(b.date_create) = :month",
             nativeQuery = true)
-    Long getMonthlySalesRevenue(@Param("year") int year, @Param("month") int month);
+    Double getMonthlySalesRevenue(@Param("year") int year, @Param("month") int month);
     @Query(value = "select day(b.date_create) as day, sum(bi.quantity * p.price) as daily_revenue " +
             "from bill_items as bi " +
             "inner join bills as b on bi.bill_id = b.bill_id " +
