+++ src/main/java/com/codegym/fashionshop/service/bill/IBillService.java
@@ -11,7 +11,7 @@ import java.util.Map;
 
 public interface IBillService extends IGeneralService<Bill> {
     boolean isBillCodeUnique(String billCode);
-    Long getDailySalesRevenue(LocalDate date);
-    Long getMonthlySalesRevenue(YearMonth yearMonth);
-    Map<Integer, Long> getDailySalesRevenueForMonth(YearMonth yearMonth);
+    Double getDailySalesRevenue(LocalDate date);
+    Double getMonthlySalesRevenue(YearMonth yearMonth);
+    Map<Integer, Double> getDailySalesRevenueForMonth(YearMonth yearMonth);
 }
